+++ tests/test_estimators/test_loess.py
@@ -1,12 +1,14 @@
 import numpy as np
 import pytest
 from sklearn.pipeline import Pipeline
-from sklearn.preprocessing import Normalizer
 
-from sklego.loess import LoessRegressor
+from sklego.linear import LoessRegressor
 
 
 def test_loessregressor_init_normal():
+    """
+    Checks initializing the LoessRegressor in normal use cases
+    """
     xs = np.linspace(start=0, stop=9, num=10).reshape(-1, 1)
     ys = np.random.random(size=10)
 
@@ -15,6 +17,9 @@ def test_loessregressor_init_normal():
 
 
 def test_loessregressor_init_errors():
+    """
+    Checks initializing the LoessRegressor with bad settings throw ValueErrors.
+    """
     xs = np.linspace(start=0, stop=9, num=10).reshape(-1, 1)
     ys = np.random.random(size=10)
 
@@ -23,8 +28,14 @@ def test_loessregressor_init_errors():
     with pytest.raises(ValueError):
         LoessRegressor(span=2).fit(xs, ys)
 
+    with pytest.raises(ValueError):
+        LoessRegressor(span=0.5, weighting_method='test').fit(xs, ys)
+
 
 def test_loessregressor_fit_normal():
+    """
+    Checks fitting the LoessRegressor returns expected values in normal use cases
+    """
     x = np.random.random(size=10).reshape(-1, 1)
     y = np.random.random(size=10)
 
@@ -34,6 +45,9 @@ def test_loessregressor_fit_normal():
 
 
 def test_loessregressor_fit_errors():
+    """
+    Checks fitting the LoessRegressor with bad values throw ValueErrors.
+    """
     with pytest.raises(ValueError):
         x = np.array([]).reshape(-1, 1)
         y = np.random.random(size=10)
@@ -51,6 +65,9 @@ def test_loessregressor_fit_errors():
 
 
 def test_loessregressor_get_window_indices_normal():
+    """
+    Checks the window indices are returned as expected.
+    """
     xs = np.linspace(start=0, stop=9, num=10).reshape(-1, 1)
     ys = np.random.random(size=10)
 
@@ -83,14 +100,21 @@ def test_loessregressor_get_window_indices_normal():
 
 
 def test_loessregressor_in_pipeline():
-    transformer = Normalizer()
-    model = LoessRegressor(span=1)
+    """
+    Check the LoessRegressor doesn't break an sklearn pipeline object in calling fit and predict
+    methods and returns expected values
+    """
 
-    pipeline = Pipeline([('normal', transformer),
-                         ('loess', model)
-                         ])
+    model = LoessRegressor(span=0.5)
+
+    pipeline = Pipeline([('loess', model)])
 
     xs = np.linspace(start=0, stop=9, num=10).reshape(-1, 1)
-    ys = np.random.random(size=10)
+    ys = [x * 2 for x in xs]
 
-    pipeline.fit(xs, ys).predict(xs)
+    y_preds_expected = [0.0, 2.0, 4.0, 6.0, 8.0, 10.0, 12.0, 14.0, 16.0, 18.0]
+
+    y_preds = pipeline.fit(xs.reshape(-1, 1), ys).predict(xs.reshape(-1, 1))
+    y_preds = [np.round(y, decimals=3) for y in y_preds]
+
+    assert y_preds == y_preds_expected
